+++ modules/azure-app-gateway/public_ip.tf
@@ -5,7 +5,5 @@ resource "azurerm_public_ip" "public_ip" {
   resource_group_name = var.resource_group_name
   sku                 = var.public_ip.sku
   allocation_method   = var.public_ip.allocation_method
-  lifecycle {
-    ignore_changes = [tags]
-  }
+  tags                = local.tags
 }