+++ modules/azure-sa-backup/shares.tf
@@ -9,7 +9,7 @@ resource "azurerm_recovery_services_vault" "this" {
   soft_delete_enabled          = var.backup_share.soft_delete_enabled
   storage_mode_type            = var.backup_share.storage_mode_type
   cross_region_restore_enabled = var.backup_share.cross_region_restore_enabled
-  tags                         = data.azurerm_resource_group.this.tags
+  tags                         = local.tags
   dynamic "identity" {
     for_each = var.backup_share.identity != null ? [1] : []
     content {
@@ -25,9 +25,6 @@ resource "azurerm_recovery_services_vault" "this" {
       user_assigned_identity_id         = var.backup_share.encryption.user_assigned_identity_id
       use_system_assigned_identity      = var.backup_share.encryption.use_system_assigned_identity
     }
-  }
-  lifecycle {
-    ignore_changes = [tags]
   }
 }