+++ modules/aws-eks/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.4.0](https://github.com/prefapp/tfm/compare/aws-eks-v1.3.1...aws-eks-v1.4.0) (2025-12-16)
+
+
+### Features
+
+* **aws-eks:** added cloudwatch_log_group_class and option for disable cloudwatch ([#780](https://github.com/prefapp/tfm/issues/780)) ([c6e6d11](https://github.com/prefapp/tfm/commit/c6e6d113c9cc8f3ba79996ff52c9e4d9e70984d7))
+
+## [1.3.1](https://github.com/prefapp/tfm/compare/aws-eks-v1.3.0...aws-eks-v1.3.1) (2025-12-11)
+
+
+### Bug Fixes
+
+* **aws-eks:** update doc config ([#775](https://github.com/prefapp/tfm/issues/775)) ([ee43ca1](https://github.com/prefapp/tfm/commit/ee43ca13c959971b29adf6f09bd81674cc58ba43))
+
+## [1.3.0](https://github.com/prefapp/tfm/compare/aws-eks-v1.2.7...aws-eks-v1.3.0) (2025-12-11)
+
+
+### Features
+
+* add karpenter ([#746](https://github.com/prefapp/tfm/issues/746)) ([591ef8a](https://github.com/prefapp/tfm/commit/591ef8a33b49b5f619a9fdfe87ec4a2abfc59745))
+
 ## [1.2.7](https://github.com/prefapp/tfm/compare/aws-eks-v1.2.6...aws-eks-v1.2.7) (2025-11-11)