+++ modules/azure-app-gateway/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.8.1](https://github.com/prefapp/tfm/compare/azure-app-gateway-v0.8.0...azure-app-gateway-v0.8.1) (2025-12-03)
+
+
+### Bug Fixes
+
+* add tags_from_rg ([#749](https://github.com/prefapp/tfm/issues/749)) ([0975dcc](https://github.com/prefapp/tfm/commit/0975dccf107749b82c9074d8bd14cd6593b2abb3))
+
+## [0.8.0](https://github.com/prefapp/tfm/compare/azure-app-gateway-v0.7.1...azure-app-gateway-v0.8.0) (2025-11-13)
+
+
+### Features
+
+* Add definition of rewrite_rule_sets for Azure App Gateway module ([#728](https://github.com/prefapp/tfm/issues/728)) ([afeaf4c](https://github.com/prefapp/tfm/commit/afeaf4c41bde6c0806e9c4143521f39f008341cb))
+
+## [0.7.1](https://github.com/prefapp/tfm/compare/azure-app-gateway-v0.7.0...azure-app-gateway-v0.7.1) (2025-11-13)
+
+
+### Bug Fixes
+
+* Solve multiple ssl-profiles definition ([#724](https://github.com/prefapp/tfm/issues/724)) ([c645dba](https://github.com/prefapp/tfm/commit/c645dba19e23ef71de1330047aa275161a360d43))
+
 ## [0.7.0](https://github.com/prefapp/tfm/compare/azure-app-gateway-v0.6.0...azure-app-gateway-v0.7.0) (2025-10-31)