+++ modules/azure-disks-backup/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.2.3](https://github.com/prefapp/tfm/compare/azure-disks-backup-v1.2.2...azure-disks-backup-v1.2.3) (2025-12-29)
+
+
+### Bug Fixes
+
+* Update README.md ([#811](https://github.com/prefapp/tfm/issues/811)) ([cb9e9c5](https://github.com/prefapp/tfm/commit/cb9e9c5936e7eeeca2c1425ed7094729b6c59089))
+
+## [1.2.2](https://github.com/prefapp/tfm/compare/azure-disks-backup-v1.2.1...azure-disks-backup-v1.2.2) (2025-12-23)
+
+
+### Bug Fixes
+
+* add tags inheritance ([#809](https://github.com/prefapp/tfm/issues/809)) ([0ae24c1](https://github.com/prefapp/tfm/commit/0ae24c12cf81f8797333f7231b5315a0369e9211))
+
 ## [1.2.1](https://github.com/prefapp/tfm/compare/azure-disks-backup-v1.2.0...azure-disks-backup-v1.2.1) (2025-04-24)